service 구현

+++ src/main/java/com/team5_jsb/domain/answer/answer/repository/AnswerRepository.java
@@ -3,5 +3,5 @@ package com.team5_jsb.domain.answer.answer.repository;
 import com.team5_jsb.domain.answer.answer.entity.Answer;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AnswerRepository extends JpaRepository<Answer, Integer> {
+public interface AnswerRepository extends JpaRepository<Answer, Long> {
 }
